# client/packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "JobEnqueueRequest": {
        "description": "Yêu cầu xếp một tác vụ vào hàng đợi.",
        "properties": {
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "JobEnqueueRequest",
        "type": "object"
      },
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobResponse": {
        "description": "Trạng thái một tác vụ, đủ để màn hình vẽ thanh tiến độ.",
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "resume_semantics": {
            "title": "Resume Semantics",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "progress",
          "message",
          "result",
          "branch_id",
          "requested_by",
          "attempt",
          "cancel_requested",
          "resume_semantics",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "done",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobTypeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobTypeResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobTypeListResponse",
        "type": "object"
      },
      "JobTypeResponse": {
        "description": "Một loại tác vụ mà bản cài này biết chạy.\n\nLộ ra để màn hình dựng được danh sách \"chạy tác vụ\" từ đúng một lượt gọi,\nthay vì mang một bản sao của registry trong client — bản sao sẽ lệch ngay ở\nphase sau, khi mỗi phân hệ thêm loại của mình.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "permission": {
            "title": "Permission",
            "type": "string"
          },
          "requires_privileged_connection": {
            "title": "Requires Privileged Connection",
            "type": "boolean"
          },
          "resume_semantics": {
            "title": "Resume Semantics",
            "type": "string"
          }
        },
        "required": [
          "code",
          "permission",
          "resume_semantics",
          "description",
          "requires_privileged_connection"
        ],
        "title": "JobTypeResponse",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "Thân lỗi RFC 7807 — **hợp đồng công khai** với client.\n\nKhai thành model (dù handler vẫn dựng `dict` để mang thêm trường tùy lớp\nlỗi) vì đúng một lý do: nó phải có mặt trong đặc tả OpenAPI, và từ đó có mặt\ntrong type sinh cho client (bước 14). Không khai thì client bắt lỗi bằng\n`any` — tức là phần *hay xảy ra nhất* của một API lại là phần duy nhất không\ncó kiểu.\n\n`model_config` cho phép trường thừa: mỗi lớp lỗi được khai thêm trường riêng\nqua `problem_extra()` (`latest` của xung đột phiên bản, `errors` của lỗi\nkiểm dữ liệu). Chúng là phần **mở rộng** của hợp đồng, không phải phần vi\nphạm nó.",
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "latest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "error_code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Konek Két — App Server",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "description": "Hàng đợi của dữ liệu kế toán đang mở, mới nhất trước.\n\nRLS lọc theo chi nhánh trước khi mã này chạy, nên \"mọi job\" ở đây nghĩa là\nmọi job **người gọi được thấy**.\n\nBộ lọc trạng thái khai bằng `JobStatus` để FastAPI trả `422` cho giá trị lạ.\nNhận `str` trần thì `?status=runing` cho ra danh sách rỗng — client đọc\nthành \"hàng đợi trống\" và không có gì báo là đã gõ sai.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Lỗi (RFC 7807)"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Xếp một tác vụ vào hàng đợi của dữ liệu kế toán đang mở.\n\n`202` chứ không `201`: yêu cầu đã được ghi nhận, kết quả chưa có. Client\ntheo dõi bằng `GET /api/v1/jobs/{id}`.\n\nQuyền riêng của loại tác vụ kiểm **trong thân hàm** vì nó phụ thuộc thân\nrequest; dependency ở chữ ký giữ mức sàn, nên người không có quyền dùng hàng\nđợi dừng lại từ trước.\n\nChi nhánh của job = chi nhánh **đang thao tác** của người yêu cầu. Đó là thứ\nquyết định phạm vi RLS mà thân job chạy dưới, nên nó phải là chi nhánh của\nngười bấm nút chứ không phải một tham số client gửi lên.",
        "operationId": "enqueue_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobEnqueueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Lỗi (RFC 7807)"
          }
        },
        "summary": "Enqueue Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/types": {
      "get": {
        "description": "Loại tác vụ mà **bản cài này** biết chạy.\n\nLọc theo quyền của người gọi: danh sách trả về là những việc họ bấm được,\nnên màn hình không hiện một nút chắc chắn trả `403`.",
        "operationId": "list_job_types_api_v1_jobs_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobTypeListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Lỗi (RFC 7807)"
          }
        },
        "summary": "List Job Types",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "Trạng thái + tiến độ của một tác vụ.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Lỗi (RFC 7807)"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Yêu cầu dừng một tác vụ.\n\nĐặt cờ chứ không giết tiến trình: worker dừng ở **ranh giới lô** gần nhất.\nGiết giữa một bút toán phân bổ là cách tạo ra sổ lệch, nên phản hồi ở đây là\n\"đã ghi nhận\", không phải \"đã dừng\" — trạng thái thật đọc ở lần `GET` sau.",
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Lỗi (RFC 7807)"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}